Move chain to util/chain. Probably shouldn't be public API.
