+++ src/util/chain.js
